+++ controller/server/update.ml
@@ -56,7 +56,8 @@ type state =
 [@@deriving sexp_of]
 
 type config =
-  { error_backoff_duration : sleep_duration
+  { http_error_backoff_duration : sleep_duration
+  ; install_error_backoff_duration : sleep_duration
   ; check_for_updates_interval : sleep_duration
   }
 
@@ -183,7 +184,7 @@ module Make (Deps : ServiceDeps) : UpdateService = struct
               )
             in
             return
-              { process_state = Sleeping config.error_backoff_duration
+              { process_state = Sleeping config.http_error_backoff_duration
               ; (* unsetting version_info to indicate we are unclear about
                    current system state *)
                 version_info = None
@@ -207,40 +208,47 @@ module Make (Deps : ServiceDeps) : UpdateService = struct
             in
             return
               { state with
-                process_state = Sleeping config.error_backoff_duration
+                process_state = Sleeping config.http_error_backoff_duration
               ; system_status = UpdateError (ErrorDownloading exn_str)
               }
       )
-    | Installing bundle_path -> (
-        (* install bundle via RAUC *)
-        match%lwt Lwt_result.catch (fun () -> RaucI.install bundle_path) with
-        | Ok () ->
-            let%lwt () =
-              Logs_lwt.info (fun m ->
-                  m "succesfully installed update (%s)" bundle_path
-              )
-            in
-            return
-              { state with
-                (* unsetting version_info, because it is now stale *)
-                version_info = None
-              ; (* going back to GettingVersionInfo to update version_info *)
-                process_state = GettingVersionInfo
-              }
-        | Error exn ->
-            let () = try Sys.remove bundle_path with _ -> () in
-            let exn_str = Printexc.to_string exn in
-            let%lwt () =
-              Logs_lwt.err ~src:log_src (fun m ->
-                  m "failed to install RAUC bundle: %s" exn_str
-              )
-            in
-            return
-              { state with
-                process_state = Sleeping config.error_backoff_duration
-              ; system_status = UpdateError (ErrorInstalling exn_str)
-              }
-      )
+    | Installing bundle_path ->
+        Lwt.finalize
+          (fun () ->
+            (* install bundle via RAUC *)
+            match%lwt
+              Lwt_result.catch (fun () -> RaucI.install bundle_path)
+            with
+            | Ok () ->
+                let%lwt () =
+                  Logs_lwt.info (fun m ->
+                      m "succesfully installed update (%s)" bundle_path
+                  )
+                in
+                return
+                  { state with
+                    (* unsetting version_info, because it is now stale *)
+                    version_info = None
+                  ; (* going back to GettingVersionInfo to update version_info *)
+                    process_state = GettingVersionInfo
+                  }
+            | Error exn ->
+                let exn_str = Printexc.to_string exn in
+                let%lwt () =
+                  Logs_lwt.err ~src:log_src (fun m ->
+                      m "failed to install RAUC bundle: %s" exn_str
+                  )
+                in
+                return
+                  { state with
+                    process_state =
+                      Sleeping config.install_error_backoff_duration
+                  ; system_status = UpdateError (ErrorInstalling exn_str)
+                  }
+          )
+          (fun () ->
+            try%lwt Lwt_unix.unlink bundle_path with _ -> Lwt.return ()
+          )
 
   let rec run_rec set_state state =
     let%lwt next_state = run_step state in
@@ -251,7 +259,8 @@ module Make (Deps : ServiceDeps) : UpdateService = struct
 end
 
 let default_config : config =
-  { error_backoff_duration = 30.0
+  { http_error_backoff_duration = 30.0
+  ; install_error_backoff_duration = 5. *. 60.
   ; check_for_updates_interval = 1. *. 60. *. 60.
   }