+++ controller/tests/server/update/update_tests.ml
@@ -55,6 +55,13 @@ let both_out_of_date ({ update_client; rauc } : Helpers.test_context) =
             in
             Lwt.return true
         )
+    ; Scenario.ActionDone
+        ( "bundle file was deleted after successful installation"
+        , fun { process_state = Installing path; _ } ->
+            Alcotest.(check bool)
+              "File no longer exists at path" false (Sys.file_exists path) ;
+            Lwt.return true
+        )
     ; Scenario.StateReached
         { base_expected_state with
           version_info = None
@@ -122,7 +129,7 @@ let delete_downloaded_bundle_on_err
     ; Scenario.StateReached
         { init_state with
           process_state =
-            Sleeping Helpers.default_test_config.error_backoff_duration
+            Sleeping Helpers.default_test_config.install_error_backoff_duration
         ; system_status = UpdateError (ErrorInstalling Scenario._WILDCARD_PAT)
         }
     ; Scenario.StateReached
@@ -146,7 +153,7 @@ let sleep_on_get_version_err _ () =
     ; system_status =
         UpdateError (ErrorGettingVersionInfo Scenario._WILDCARD_PAT)
     ; process_state =
-        Sleeping Helpers.default_test_config.error_backoff_duration
+        Sleeping Helpers.default_test_config.http_error_backoff_duration
     }
   in
   let%lwt out_state = UpdateServiceI.run_step init_state in
@@ -171,7 +178,7 @@ let sleep_on_download_err _ () =
     { version_info = None
     ; system_status = UpdateError (ErrorDownloading Scenario._WILDCARD_PAT)
     ; process_state =
-        Sleeping Helpers.default_test_config.error_backoff_duration
+        Sleeping Helpers.default_test_config.http_error_backoff_duration
     }
   in
   let%lwt out_state = UpdateServiceI.run_step init_state in